Re-factoring for use with upstream Xen.
Strip out XenServer specific parts of the drivers and enable them
to cope with multiple bindings.
Also, refine some of the source, add a common source area to reduce
code duplication, and change the PDO names to reflect the PV driver
vendor and also use a format that should work with Windows Update.

Note that XENFILT is pretty useless at this point. It binds to devices
correctly but all the unplug code is gone, to be replaced with code that
is not reliant upon XenServer-specific patches to QEMU.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>