Modify GNTTAB interface to support caches
Remove the Get/Set methods and replace with CreateCache/DestroyCache methods.
The PermitForeignAccess method the uses a cache to allocate an entry as it
requires. Also, as we're not only using grant table v1, we can get rid of
the idea of different types of permit and drop the use of varargs.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>